Persist scheduled provider sync runs
