+++ tests/unit/core-api/ingestion-scheduler.test.ts
@@ -520,7 +520,11 @@ describe('IngestionScheduler', () => {
         },
       );
 
-      await (scheduler as any).runConfiguredSportScheduleSync('GOLF' as Sport);
+      const runConfiguredSportScheduleSync = Reflect.get(
+        scheduler,
+        'runConfiguredSportScheduleSync',
+      ) as (sport: Sport) => Promise<void>;
+      await runConfiguredSportScheduleSync.call(scheduler, 'GOLF' as Sport);
       await (scheduler as any).runConfiguredSportFieldSync('GOLF' as Sport);
       await (scheduler as any).runConfiguredSportRankingSync('GOLF' as Sport);
 
@@ -554,6 +558,77 @@ describe('IngestionScheduler', () => {
         },
       });
       expect(mockCallbacks.onJobComplete).toHaveBeenCalledTimes(3);
+    });
+
+    it('pool-master-rop.68.2.4 records configured sport syncs in the provider sync run ledger once per ingestion job', async () => {
+      const now = new Date('2026-04-28T12:00:00.000Z');
+      const provider = createMockProvider({
+        getUpcomingEvents: jest.fn().mockResolvedValue([]),
+      });
+      const config = createEnabledScheduleConfig();
+      const configReader = {
+        getConfig: jest.fn().mockResolvedValue(config),
+        getPerSportConfig: jest.fn().mockResolvedValue(config),
+      };
+      const syncRun = {
+        id: 'scheduled-sync-run-1',
+        providerId: 'mock-provider',
+        sport: 'GOLF' as Sport,
+        eventId: null,
+        status: 'SUBMITTED' as const,
+        startedAt: null,
+        completedAt: null,
+        createdAt: now,
+        payload: {
+          requestedFeed: 'EVENTSCHEDULE',
+        },
+      };
+      const syncRunLedger = {
+        createSubmissions: jest.fn().mockResolvedValue([syncRun]),
+        executeFeedRun: jest.fn(async (_syncRun: typeof syncRun, run: () => Promise<unknown>) => {
+          const job = await run();
+          return job as Awaited<ReturnType<IngestionScheduler['syncSport']>>;
+        }),
+      };
+      const scheduler = new IngestionScheduler(
+        createMockRegistry(provider, ['GOLF' as Sport]),
+        mockCallbacks,
+        undefined,
+        {
+          configReader,
+          now: () => now,
+          syncRunLedger,
+        },
+      );
+
+      const runConfiguredSportScheduleSync = Reflect.get(
+        scheduler,
+        'runConfiguredSportScheduleSync',
+      ) as (sport: Sport) => Promise<void>;
+      await runConfiguredSportScheduleSync.call(scheduler, 'GOLF' as Sport);
+
+      expect(syncRunLedger.createSubmissions).toHaveBeenCalledWith(expect.objectContaining({
+        providerId: 'mock-provider',
+        runType: 'SCHEDULED_SPORT_SYNC',
+        submittedAt: now,
+        normalizedRequest: expect.objectContaining({
+          source: 'SCHEDULED',
+          actor: { type: 'SYSTEM', name: 'scheduler' },
+          scope: expect.objectContaining({
+            type: 'SPORT',
+            sport: 'GOLF',
+            feeds: ['EVENTSCHEDULE'],
+            effectiveWindow: {
+              from: now,
+              to: new Date('2026-05-28T12:00:00.000Z'),
+              defaultedFrom: true,
+              defaultedTo: true,
+            },
+          }),
+        }),
+      }));
+      expect(syncRunLedger.executeFeedRun).toHaveBeenCalledWith(syncRun, expect.any(Function));
+      expect(mockCallbacks.onJobComplete).toHaveBeenCalledTimes(1);
     });
 
     it('pool-master-rop.68.2.2 submits configured event loops as scheduled system sync requests', async () => {
